Ensure controlapi Server always has an up to date allocator
Currently the allocator is only created when a node becomes the leader, which
is (almost always) after the controlapi.NewServer call. Hence the server is
always being passed a m.allocator which is nil.

In addition the server wasn't actually remembering the allocator it was passed.

Add a second layer of indirection to the allocator stashed by the controlapi
Server, so it will always see the latest version of m.allocator.

This stops #2230 from happening, but is really only a stopgap until a more
satisfactory solution can be found.

This regression was made by me in fb74191488a7 ("Convert NetworkAllocator to an
interface"). Add a call to CreateNetwork to TestManager intended to exercise
the code path through validateNetwork to allocator.IsBuiltinNetworkDriver.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>