scheduler: Do not try to underweight nodes for global services
The scheduler has simple failure loop detection that attempts to prefer
other nodes when a certain node experiences several failures in a row of
tasks from the same service. This makes sense for replicated services,
but doesn't have an effect on global services, because the node is
predetermined for each task. When this code path triggers with a global
service, it produces a confusing log message that suggests other nodes
will be preferred for this global service.

Change the scheduler to keep track of which tasks were preassigned to
nodes, and avoid tracking failures of those tasks.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>