orchestrator: Check that service's SpecVersion exists before dereferencing
IsTaskDirty assumed that if a task had a SpecVersion, the corresponding
service must also have a SpecVersion, since that is where the task's
field is copied from. However, there are mixed-version scenarios where a
service could get updated by a manager running an older version of
swarmkit than the manager that created the task. Add a check to avoid a
crash in this case.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>