Avoid nil Allocator in controlapi
At the point at which the controlapi server is created the allocator has not
yet been created, and hence is nil.

Since CNM is currently the only actual choice do as with `PredefinedNetworks`
and hardcode the use of the CNM backend, this will be reworked once there is a
second choice.

Fixes #2230. Add a call to NetworkCreate in TestManager and a test to
controlapi specifically to trigger this code path. The latter tests controlapi
in isolation, while the former would catch e.g. a race or initialisation order
issue in a manager.

Also, in controlapi tests, change a `testnet2` into a 3 in `TestRemoveNetwork`
to avoid (harmless) duplication.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>