Fixing error with host network mode on upgrade from pre 17.06 versions
This commit contains the following fixes:
1) On upgrade from pre 17.06 version services created with net=host
were failing because the swarm was not aware of host network being
created. The fix added takes care of this by creating network if
the store does not contain predefined networks

2) On a cluster with upgraded 17.06 version of docker swarm manager
the daemon panics while trying to access IPAM driver in the createNetworkRequest on
the nodes with pre 17.06 versions. In order to avoid panics empty struct is passed
with network attachment. This does not have any function impact.

Signed-off-by: Abhinandan Prativadi <abhi@docker.com>