Skip TestUpdaterRollbackAndPauseNoMonitor
This test triggers a race condition in the way the updater and restart
supervisor interact.

The problem is normally masked by a second reconciliation pass triggered
as a side effect of the update completing, but this test case does not
wait long enough for the update to be declared complete.

Since the test does not pass reliably, it should be skipped until the
underlying issue is fixed.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>