scheduler: Version-aware failure tracking
This addresses a longstanding TODO which I forgot about. Currently, the
scheduler tracks potential failure loops by service ID. However, if the
service is updated, it won't distinguish the new version of the service
from the old one. This means even if the operator fixes the underlying
problem, the new tasks from the updated service will still avoid the
node where the problem was detected.

Change the failure tracking to be sensitive to SpecVersion as well. If
SpecVersion changes, we'll treat the task as a different kind of task
for failure tracking purposes.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>