scheduler: Clean up addOrUpdateNode
This function previously could take an uninitialized NodeInfo structure
and fill in whatever was missing. This is very error-prone, so remove
this logic and change the only caller that relies on it to always pass
in a properly initialized NodeInfo.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit 5b4b55a2490f19b0507b0aae28116e74720598ac)