Run CI with temporary files on a ramdisk
This sets up a tmpfs mount to be used for temporary files created by tests.

If this works, hopefully it will fix the excessive I/O latencies that
are blocking tests for long periods and causing them to fail.

As a side effect, this also stores temporary compilation artifacts on
tmpfs.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>