template: Indicate which configs contain secret data
This changes the templated ConfigGetter to include an extra method,
GetAndFlagSecretData, which returns an interpolated config and also a
flag indicating whether it contains any data from a secret. This can be
used downstream to avoid storing the config to disk.

Ideally, there would be some kind of Sensitive method on a Config object
itself, but since that type is automatically generated by protoc,
including a private field is awkward to do. Thus this approach of
returning the flag along with the config. Note that I opted for a method
that returns the config and the flag at the same time instead of a
method that just returns the flag, since a method that returns the flag
still needs to expand the config, and it seems most robust and
performant to just return the config and the flag at the same time
instead of relying on different code paths to generate consistent
results.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>