api: Use TaskStatus Err field for non-terminal errors
There are some cases when a task can't advance from a particular state
because preconditions are not met. For example, if no nodes meet its
constraints, it will not advance to "assigned".

Currently, we put a note about this in the Message field of TaskStatus,
but this is not surfaced to the user. It wouldn't make sense to expose
Message prominently because it usually contains uninteresting notes
about how the task reached that state. Its presence does not indicate
that something is wrong.

Expand the scope of Err to also cover non-terminal errors that are
blocking the task from progressing, and use it in those cases.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>