Respect restart policy during service reconciliation
Currently, service reconciliation (updating a service, leader
reelection, etc) will cause a service's tasks to be restarted even if
the restart policy says they shouldn't be. This is because the service
definition says there should be, for example, 5 replicas, and the
control loop tries to make this happen.

This experiments with the idea of keeping tasks that we don't want to
restart in the desired state "running" instead of "shutdown". This can
be a starting point for evaluating the tradeoffs and deciding if this is
a good approach.

Known caveats so far:

- Draining a node has to set the desired state of tasks on that node to
  "shutdown", regardless of the restart policy. But that means the
  orchestrator may eventually restart those tasks. So draining triggers
  the same issue we had before.
- When rolling back a partial service update, tasks with the original
  service definition that happen to have failed won't be restarted. I'm
  not sure this is the correct behavior, because I would expect a service
  rollback to leave the service in a converged state.
- Global services have tasks shut down on nodes that go down. This is
  necessary to prevent rolling updates from getting stuck on those
  nodes. But it means that a node going down and coming back up will
  cause a restart of its global service tasks, even if that does not
  match the restart policy.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>