Push network spec validation down into network.Model
Also add a SetDefaults which avoids updating the spec in the validation
function in the CNI case.

With this all uses of libnetwork are in manager/allocator/cnmallocator and
manager/network/cnm/cnm.go while all uses of CNI are in
manager/network/cni/cni.go and agent/exec/containerd/adapter.go. I'd still like
to abstract the containerd/adaptor.go case further.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>