orchestrator/update: Only shut down old tasks on success
In start-then-stop mode, the updater waits for a verdict on the startup
of a new task, and then shuts down the tasks that this new task was
meant to replace. It should really only shut down the old task if the
new task successfully reached running. The thinking here was probably
that a task which fails on startup would briefly enter the running state
before failing, so there isn't much point in distinguishing between
running and some terminal state. However, if a health check is being
used, the task may never enter the running state. This can also happen
when a task is rejected by a node.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>