orchestrator/update: Update spec version when rolling back
There is an oversight in this code that prevents it from changing the
spec version to match the spec that's put in place during a rollback.

This could cause a service not to get rolled back all the way.
Currently, tasks may stay on the new version until they fail and are
replaced with an older version.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>