watchapi: Terminate RPCs when the manager shuts down
Because we use GracefulStop, RPCs need to finish on their own when the
manager shuts down.

The watchapi server formerly was completely stateless, and didn't have a
notion of starting or stopping.

This adds Start and Stop functions that control a context, adapted from
the logbroker code. When the context is cancelled, outstanding RPCs will
finish.

There are also some related cleanups and fixes in the logbroker code.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>