manager: Always run the watch server
The watch server was wrongly changed to only run on the leader node. It
needs to run on all managers because this is one of the RPC services
that is not proxied to the leader (since all nodes receive events
through Raft).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit 80abd938845447910a6093ee7b3501074144cf91)