orchestrator: Use task history to evaluate restart policy
Previously, restart conditions other than "OnAny" were honored on a
best-effort basis. A service-level reconciliation, for example after a
leader election, would see that not enough tasks were running, and start
replacement tasks regardless of the restart policy. This limited the
usefulness of the other restart conditions.

This change makes the orchestrators check historic tasks to figure out
if a task should be restarted or not, on service reconciliation.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>