Rework ContainerNetworkStatus
Structure as Network->[]Interfaces->[]IPs, which more closely reflects reality
and is also closer to the CNI model.

While reworking this I observed the CNI_IFNAME environment variable (AKA
libcni.RuntimeConf.IfName) is optional, however with at least the weave-net
plugin (and looking at the CNI repo anything based on the skeleton there)
insists on it.

Relatedly I also see that the Result lacks an Interfaces array and that the IP
addresses reference Interface==-1.

Handle this in the code by creating a fake Interface if none is returned (using
the network if name) and using it if the index is out of bounds.

Signed-off-by: Ian Campbell <ijc@docker.com>