cni: Do not override Annotations name to match CNI name
The spec should be readonly to swarmkit. Instead push this requirement onto the
user. This is already checked by `cni.ValidateNetworkSpec` but relax this into
just a warning, so admins have the flexibility to have a different CNI name to
swarm name. However log the mismatch and pass a context to
`network.Model.ValidateNetworkSpec()` to facilitate that.

Signed-off-by: Ian Campbell <ijc@docker.com>