Removing creation of predefined networks on swarm manager
From 17.06 we started supporting service creation on host and bridge
predefined networks in addition to user defined local scope networks.
Since the task allocation involves creation of networks in the swarm
manager. A network id is generated on the manager to create predefined networks.
But this network id is not visible with the api to fetch networks. Since
the tasktemplate which is part of a servicespec holds this networkid
its hard to correlate the network id to network name by the api consumer.
For this reason the service spec is created to support network target as
ID,host or bridge. The change that was introduced in 17.06 to create dummy
predefined networks is removed.With this change predefined networks are not
created on the swarm manager.This also means the allocator will be skipped
for these networks for a service.

Signed-off-by: Abhinandan Prativadi <abhi@docker.com>