vendor: Revendor
There are some files missing from the vendor directory which vndr
thinks should be there. This probably happened because of colliding
merges, or a PR that didn't add the full vendor directory. It causes
vendor validation to fail on PRs that touch the vendor directory or are
based on an old version of the tree.

Rerun vndr.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>