PoC implementation
I want to address #1895 - Please beaware that this is just an experiment
to get my hands dirty with the codebase and to understand how things
works.

At the current stage this basically is a gRPC service on the managers
that connects the client to the client to the container executing the
task of a specific service.

It does not have TTYs.

Feedback and What I need to go on:
- I need some advice in designing this and if possible code review.
- Since this is just a PoC I'm not yet writing tests, this is just
something I need to understand a codebase where I'm not even that
familiar to do TDD.
- There are probably, I already see some of them contexts that are not
closed.

Try it:
  swarmctl service exec <service-id>

(Yes I know that one expects to have `sh`) in the end or something like
that. It's too early

Signed-off-by: Lorenzo Fontana <lo@linux.com>