[WIP] - Secrets for plugin services
Work in progress
Support specifying secrets in generic tasks (e.g., plugins).
Currently, populating secrets to plugins is not supported since the
secrets references are under the container runtime struct (which is
mutually exclusive to the plugin struct).
As discuss, for backward compatibility, I've added the `SecretReference`
structs under the `Task` struct, and unioned the secret references with
the existing container struct in all code occurrences.

I will add more tests and comments if you think this is the correct
approach.

Signed-off-by: Liron Levin <liron@twistlock.com>