Delete node attachments when node is removed
When a node is removed, delete all of its attachment tasks, so that any
networks being used by those tasks can be successfully removed.

Provides a workaround to the state where a node with attachments is
somehow removed from the cluster while attached to a network, preventing
the network from being removed. Does not fix many other related bugs.

Includes a unit test for the function that removes node attachment
tasks.

Signed-off-by: Drew Erny <drew.erny@docker.com>