Addressing duplicate IP issue on ingress network
On leader change or leader reboot the restore logic in the allocator
was allocating overlapping IP address for VIPs and Endpoints in the
ingress network.  The fix added as part of this commit ensures
that during restore we allocate the existing VIP and endpoint.

Signed-off-by: Balraj Singh <balrajsingh@ieee.org>
(cherry picked from commit 5fd25d2bf65d543f0a36c0202a305d063bf443e5)