Fix task reaper not cleaning up COMPLETE tasks
Tasks in the complete state were not cleaned up because the task reaper
was only looking at task.Status.State >= api.TaskStateShutdown, but the
COMPLETE state (api.TaskStateCompleted) is the actual "first" terminal
state.

Added a test for the initialization stage of the task repear. This runs
the logic of which tasks to clean up through its paces, but doesn't test
the run-time logic of the task reaper (it's ability to respond to
updates to tasks).

Cherry-pick of docker/swarmkit#2473

Cherry pick does not apply cleanly. In addition, cherry pick requires
changes to `manager/orchestrator/taskreaper/task_reaper_test.go`,
because of a difference in the API of the task reaper.

(cherry picked from commit e5b310727c4c8b186c774b4e28d154c3fb4e32b8)
Signed-off-by: Drew Erny <drew.erny@docker.com>