WIP: Resend task statuses when reestablishing session
Oh boy this one is a doozy

1. Makes the status reporter support batch updates
2. Makes the agent support batch updates as a status reporter
3. Makes the places that use a status report use batch updates when
appropriate
4. Add's a `Report` method to the worker, which reports the status of
all tasks to the given StatusReporter
5. Makes the agent call the worker's `Report` method every time it
reestablishes a session.

This fixes a race condition that results from the Dispatcher in the
manager batching updates, and telling the agent that task updates have
succeeded when they haven't actually been committed to the raft log.
This has the effect of making the managers drop status updates if there
is a leadership change while a batch of status updates is in the
pipeline. The fix here is that when agent's reestablish a session, they
re-send all of their task statuses, so the manager will DEFINITELY be
up to date.

Signed-off-by: Drew Erny <drew.erny@docker.com>