Avoid predefined error log
- fix the logic that was checking for predefined networks. Before the
GetNetwork was passing the network name, while instead the network ID is
the key that is supposed to be used. The CreateNetwork anyway is doing
the very same check and create the network if not existing.

Signed-off-by: Flavio Crisciani <flavio.crisciani@docker.com>
(cherry picked from commit f6e6e694ed86f6ca470bf67112722eb5ade17bf5)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>