If the cluster requires FIPS, the cluster ID and join token will also reflect this property. So all TLS certs will have the cluster ID, which says whether the cluster is FIPS, in the Org field.
If a node loads up its TLS cert, sees that that the cluster requires FIPS,
and FIPS mode is not enabled on that node, the node will shut down.  If a
non-FIPS node gets a join token that indicate that the cluster mandates
FIPS, it will refuse to join.

Signed-off-by: Ying Li <ying.li@docker.com>
(cherry picked from commit 9943770cf2e42aef706b8a1082cba33a911752f9)