Fix some flakiness with the manager tests:  in the first case (TestManager), it may take a little while for the dispatcher to come up, so poll until it's up.
In the second case (TestManagerLockUnlock), we wait for the TLS key bytes to change to
test to see whether it has been decrypted after disabling the unlock key.  However, it
could have changed due to a renewal due to the previous unlock key rotation.  So when
polling, check that it's decrypted.

Signed-off-by: Ying Li <ying.li@docker.com>