update the task status to failed if the task status is ready or staring when creating new taskmanager of the task
When creating a new taskmanager, setting the task status to failed if the task status is ready or starting,so that the leader shutdown the task and recreate a task.

The container's NetworkSettings.Service is assigned by calling ctlr.Prepare when the task status.State is api.TaskStatePreparing, but don't  save it to tasks.db or
related disk of the container.
when the task state is api.TaskStateReady or api.TaskStateStarting, the host power down or restarts abnormally, and the container's NetworkSettings.Service is lost.

When the agent init worker, the task state restores from the previous state, the container starts to run normally, but the NetworkSettings.Service information is lost,
there is no iptable rules and  the port of ingress network.

This problem is occasional,but we can add sleep when the task status is ready and power down the host of agent side.
(https://github.com/docker/swarmkit/blob/master/agent/exec/controller.go:L343)

Signed-off-by: ZhiPeng Lu <lu.zhipeng@zte.com.cn>