Add some TLA+ specifications for SwarmKit
These documents use the TLA+ notation to describe the behaviour of
SwarmKit more precisely than the prose documentation in the design
directory.

Run `make pdfs` to render these documents as PDF files.
The best one to start with is `SwarmKit.pdf`, which introduces the TLA+
notation and describes the overall components of SwarmKit.

The specifications can also be executed by the TLC model checker to help
find mistakes. Use `make check` to run the checks.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>