This commit fixes some issues with our use of tls.Config and with GRPC:
1. We now call `Clone` on the tls config when cloning the MutableTLSConfig
because you can't reuse a tls.Config for multiple connections, which we
would be doing if something called MutableTLSCreds.Clone, which newer
versions of GRPC do.

2. To prevent data races when grpc attempts to log, only set the grpc logger
to discard in test init functions, rather than in TestMain functions as is
documented by the `SetLoggerV2` function.  Also, use `SetLoggerV2` since
SetLogger is deprecated.

3. Wrap the default logger in a private struct that converts a logrus
entry into a grpc LoggerV2, since the grpc Logger is deprecated.

4. Stop using transport.StreamFromContext to get the peer address when checking
redirects in the raft proxy.  grpc/transport is an internal package that
downstream should not depend on, as the API may change (StreamFromContext
for example is removed in >=1.11).  peer.FromContext can get the peer address
just as well.

Signed-off-by: Ying Li <ying.li@docker.com>