Disallow deleting networks that are in use
Change the control API to disallow removing networks that are in use by
tasks. Before, if a task had a terminal desired state, the network
remove would be allowed to proceed. However, if the task is still
running, its resources are still in use. This removes the check for
desired state, so that only actual state is taken into consideration
when checking if a network is in use.

Signed-off-by: Drew Erny <drew.erny@docker.com>