Add additional info for secret drivers
This provides more context for the secret driver when it is requested
the value for the secret. It is useful both for audit purposes, e.g. an
external system logging which task requested what secret, as well as
in a scenario where the plugin would return a different value (or
error) based on e.g. labels on the secret.

Signed-off-by: Sune Keller <absukl@almbrand.dk>