Fix leaking task resources
Somehow it's possible for tasks to be deleted in between the time that
the task is allocated and the time that the task update is committed to
the store. Previously, if that happened, the network resources would
just get dropped (although they would be freed back up on leadership
change).

With this change, if we try to commit a task but find it no longer
exists, we should deallocate its resources.

Signed-off-by: Drew Erny <drew.erny@docker.com>