Adding a new `Deallocator` component
This patch adds a new manager/deallocator package.

The deallocator's job is to take over deletion of services and service-level
resources (as of right now, only networks, but e.g. volumes would also fall in
the same category.

The deallocator relies on the reaper correctly deleting tasks for services that
are shutting down, and only when no tasks are left within a service does it
then proceed to delete that service and its resources.

This patch does not yet actually put that new component to use; that will be
done in a future, separate patch, for the sake of easier reviews.

Special care has been taken to separate the deallocator itself from each of the
service-level resource implementation.

Includes unit tests.

Signed-off-by: Jean Rouge <jer329@cornell.edu>