Include old error-message for backward compatibility
Commit 2061af766f61f6bc8d3e43d5553fba315f569597 fixed the API
returning incorrect status-codes, but also changed the error
message for conflicting service-names to be in line with other
objects (secrets, configs); "service XX already exists".

Unfortunately, there are existing consumers of the API that
perform string-matching, and changing the error-message resulted
in a breaking change.

This patch prepends the `ErrNameConflict` error-message to the
error-message, so that those consumers still find the original
message, but preserves the enhancement made in 2061af7 (inclusion
of the conflicting service name).

With this patch applied, the error message will look like this;

    name conflicts with an existing object: service myservice already exists

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>